refactor: rename `EventBuffer` to `PendingEventsBuffer`, adding documentation
