+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/sync/SyncManager.kt
@@ -116,7 +116,7 @@ internal class SyncManagerImpl(
     private var processingJob: Job? = null
 
     // Do not access this variable directly of I will cut off your hands
-    private val offlineEventBuffer = EventBuffer()
+    private val offlineEventBuffer = PendingEventsBuffer()
 
     private var processingEventFlow = MutableSharedFlow<Event>()
